Edit custom CSS file
This edit removed padding and margin from the .terminal
class which caused the text in pre tags to be displayed
at the bottom.

Together with the change made to the properties in the pre tag
selector made it possible to format terminal output that
is placed inside of shortcodes correctly.