Prefix post files with post numbers
Post files were intially named using only short, descriptive names.
Prefixing the file names with a serialized number, which begins with 000 and
increments by one on each successive post, makes it easier to files in the
project project explorer.

The 'slug' front matter was added to be able to refer to posts using the old nomenaclature
in the path.

Shortcodes that reference other posts need to include the serialized prefix. For example,
{{< ref "002-sdkman-ftw" >}}